Removed a ton of msvc warnings revealed with OGRE v1.7 (they removed the warning suppressors in OgrePrerequisites.h). All of them are conversions from one type to another that might be lossy (mostly double to float, please always use "3.7f" instead of "3.7" as constants when using floats).

+++ code/trunk/src/libraries/core/GraphicsManager.cc
@@ -365,7 +365,7 @@ namespace orxonox
 
         uint64_t timeAfterTick = time.getRealMicroseconds();
         // Subtract the time used for rendering from the tick time counter
-        Game::getInstance().subtractTickTime(timeAfterTick - timeBeforeTick);
+        Game::getInstance().subtractTickTime((int32_t)(timeAfterTick - timeBeforeTick));
 
         // again, just to be sure OGRE works fine
         ogreRoot_->_fireFrameEnded(evt); // note: uses the same time as _fireFrameStarted
